requirements.txt
python-telegram-bot==20.7
spotipy==2.23.0
yt-dlp==2023.12.30
youtube-search-python==1.6.6
mutagen==1.47.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0 